Plot multiple ROC curves

Given a list of results computed by calculate_roc, plot the curve using ggplot with sensible defaults. Pass the resulting object and data to export_interactive_roc, plot_interactive_roc, or plot_journal_roc.

Usage

multi_ggroc(
  datalist,
  fpf_string = rep("FPF", length(datalist)),
  tpf_string = rep("TPF", length(datalist)),
  c_string = rep("c", length(datalist)),
  label = NULL,
  legend = TRUE,
  label.adj.x = rep(0, length(datalist)),
  label.adj.y = rep(0, length(datalist)),
  label.angle = rep(45, length(datalist)),
  plotmath = FALSE,
  xlabel = "False positive fraction",
  ylabel = "True positive fraction"
)

Arguments

datalist

List of data frames each containing true and false positive fractions and cutoffs

fpf_string

Column names identifying false positive fraction

tpf_string

Column names identifying true positive fraction

c_string

Column names identifying cutoff values

label

Not supported.

legend

If true, draws legend

label.adj.x

Not supported.

label.adj.y

Not supported.

label.angle

Not supported.

plotmath

Logical. Not supported.

xlabel

Defaults to "False positive fraction"

ylabel

Defaults to "True positive fraction"

Value

A ggplot object